backend/requirements.txt
fastapi
uvicorn
pypdf2
openai
faiss-cpu
numpy
python-multipart